fixed call domain_id

+++ model/call.go
@@ -118,7 +118,7 @@ type Call struct {
 type CallAction struct {
 	Id        string `json:"id"`
 	AppId     string `json:"app_id"`
-	DomainId  int8   `json:"domain_id,string"`
+	DomainId  int64  `json:"domain_id,string"`
 	Timestamp int64  `json:"timestamp,string"`
 	Event     string `json:"event"`
 }
